# engineering/api/openapi.yaml
openapi: 3.1.0
info:
  title: AVENIQ Investigation API
  version: 0.1.0
  description: |
    Contract-first API for incident investigations (Phase 0b–1).
    Semantic model: engineering/lld/01-domain-model.md
servers:
  - url: http://127.0.0.1:8000
    description: Local development
  - url: https://api.staging.example.com
    description: VPS staging (future)
...
paths:
...
  /api/v1/playground/scenarios/{scenario_id}/runs:
    post:
      tags: [playground]
      operationId: runPlaygroundScenario
      summary: Run a Playground scenario end-to-end
      description: |
        Resets and triggers the checkout stand-in, creates an investigation, runs the playbook.
        Investigation connectors remain fixture-based in Slice 2.
      parameters:
        - name: scenario_id
          in: path
          required: true
          schema:
            type: string
            example: b03
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "201":
          description: Playground run completed
          headers:
            X-Request-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaygroundRunResponse"
        "404":
          $ref: "#/components/responses/PlaygroundScenarioNotFound"
        "502":
          $ref: "#/components/responses/PlaygroundUpstreamError"
        "504":
          $ref: "#/components/responses/PlaygroundTimeout"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
...
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
  schemas:
...
    Signal:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          enum: [alert, symptom, manual]
        fired_at:
          type: string
          format: date-time
        service:
          type: string
        description:
          type: string
        raw:
          type: object
          additionalProperties: true
    InvestigationState:
      type: string
      enum:
        - initialized
        - context_gathering
        - investigating
        - rca_candidate
        - inconclusive
        - blocked
    TimeWindow:
      type: object
      required: [start, end]
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    InvestigationHeader:
      type: object
      required:
        - id
        - benchmark_id
        - state
        - created_at
        - updated_at
        - run_count
      properties:
        id:
          type: string
        benchmark_id:
          type: string
        state:
          $ref: "#/components/schemas/InvestigationState"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        run_count:
          type: integer
          minimum: 0
        org_id:
          type: string
          nullable: true
        environment_id:
          type: string
          nullable: true
        incident_window:
          $ref: "#/components/schemas/TimeWindow"
        run_completed_at:
          type: string
          format: date-time
          nullable: true
        signal:
          $ref: "#/components/schemas/Signal"
    EvidenceStrength:
      type: string
      enum: [direct, derived, correlated, inferred, human_provided]
    EvidenceSourceType:
      type: string
      enum:
        - telemetry_log
        - telemetry_metric
        - telemetry_trace
        - change_deployment
        - context_service
        - alert
        - human
    EvidenceProvenance:
      type: object
      properties:
        connector:
          type: string
        fixture_path:
          type: string
        fixture_ref:
          type: string
        query:
          type: object
          additionalProperties: true
    Evidence:
      type: object
      required:
        - id
        - source
        - source_type
        - entity
        - observation
        - event_time
        - retrieved_at
        - strength
        - provenance
      properties:
        id:
          type: string
        source:
          type: string
        source_type:
          $ref: "#/components/schemas/EvidenceSourceType"
        entity:
          type: string
        observation:
          type: string
        event_time:
          type: string
          format: date-time
        retrieved_at:
          type: string
          format: date-time
        strength:
          $ref: "#/components/schemas/EvidenceStrength"
        provenance:
          $ref: "#/components/schemas/EvidenceProvenance"
    HypothesisStatus:
      type: string
      enum: [proposed, supported, weakened, rejected]
    Hypothesis:
      type: object
      required: [id, statement, status, supporting_evidence_ids, contradicting_evidence_ids]
      properties:
        id:
          type: string
        statement:
          type: string
        status:
          $ref: "#/components/schemas/HypothesisStatus"
        supporting_evidence_ids:
          type: array
          items:
            type: string
        contradicting_evidence_ids:
          type: array
          items:
            type: string
    RcaStatus:
      type: string
      enum: [confirmed, probable, possible, inconclusive]
    RcaClaim:
      type: object
      required: [id, text, evidence_ids]
      properties:
        id:
          type: string
        text:
          type: string
        evidence_ids:
          type: array
          items:
            type: string
          minItems: 1
    RcaTimelineEntry:
      type: object
      required: [at, description]
      properties:
        at:
          type: string
          format: date-time
        description:
          type: string
    Rca:
      type: object
      required:
        - status
        - summary
        - root_cause
        - contributing_factors
        - causal_chain
        - timeline
        - affected_components
        - claims
        - contradicting_evidence_ids
        - evidence_gaps
        - remediation
        - observability_lessons
      properties:
        status:
          $ref: "#/components/schemas/RcaStatus"
        summary:
          type: string
        root_cause:
          type: string
        contributing_factors:
          type: array
          items:
            type: string
        causal_chain:
          type: array
          items:
            type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/RcaTimelineEntry"
        affected_components:
          type: array
          items:
            type: string
        claims:
          type: array
          items:
            $ref: "#/components/schemas/RcaClaim"
        contradicting_evidence_ids:
          type: array
          items:
            type: string
        uncertainty:
          type: string
          nullable: true
        evidence_gaps:
          type: array
          items:
            type: string
        remediation:
          type: array
          items:
            type: string
        observability_lessons:
          type: array
          items:
            type: string
    InvestigationEvent:
      type: object
      required: [seq, type, occurred_at, payload]
      properties:
        seq:
          type: integer
          minimum: 1
        type:
          type: string
        occurred_at:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
...
    PlaygroundRunResponse:
      type: object
      required: [scenario_id, investigation_id, state]
      properties:
        scenario_id:
          type: string
        investigation_id:
          type: string
        state:
          type: string
        investigation:
          $ref: "#/components/schemas/InvestigationResponse"
    InvestigationResponse:
      type: object
      required: [investigation, events, evidence, hypotheses]
      properties:
        investigation:
          $ref: "#/components/schemas/InvestigationHeader"
        events:
          type: array
          items:
            $ref: "#/components/schemas/InvestigationEvent"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
        hypotheses:
          type: array
          items:
            $ref: "#/components/schemas/Hypothesis"
        rca:
          $ref: "#/components/schemas/Rca"
          nullable: true
    ProblemDetail:
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
...
    PlaygroundScenarioNotFound:
      description: Unknown Playground scenario
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://aveniq.dev/problems/scenario_not_found
            title: Not Found
            status: 404
            code: scenario_not_found
    PlaygroundUpstreamError:
      description: Stand-in or upstream HTTP failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://aveniq.dev/problems/playground_upstream_error
            title: Bad Gateway
            status: 502
            code: playground_upstream_error
    PlaygroundTimeout:
      description: Playground run timed out waiting for terminal investigation state
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://aveniq.dev/problems/playground_timeout
            title: Gateway Timeout
            status: 504
            code: playground_timeout
    InternalError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://aveniq.dev/problems/internal_error
            title: Internal Server Error
            status: 500
            code: internal_error
            detail: An unexpected error occurred
...